requirements.txt
pandas
numpy
seaborn
matplotlib
opencv-python
opencv-python-headless
pillow
scikit-learn
tensorflow

fastapi
uvicorn
tensorflow
pillow
joblib
numpy
